+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- CPX-AP: Added INT16 and UINT16 support for MPA modules in read/write functions, expanded system tests
+
 ## v0.4.2 - 29.05.24
 ### Changed
 - Updated links to public repo